Extract elementWithRelativePathMatches from ExecutionNodeAccessHierarchy
to re-use it later on.
